Adding new Browser Labs feature: Slide Tab Transitions
Edge swipe from the sides of the screen to view the previous/next tab in
the same way as the Android homescreen navigates its views. The feature
is enabled by default and can be turned off in Browser settings under
Labs. It is only enabled for the phone UI, it will not run on the tablet
interface. Slide Tab Transitions is incompatible with Quick Actions.

There is a known issue with the current patch:
- Flakiness with the title bar show/hide due to switch to non-embedded
  title bar: At times the title bar does a show/hide too quickly to
  interact with it and it requires reactivating the title bar again.

Using RealViewSwitcher.java widget from CAF with the following changes:
- Added onScrollStartedListener to notify when a View switch starts
- Added orientationChanged callback

author: Pierre-Antoine LaFayette <plafayet@codeaurora.org>
https://www.codeaurora.org/gitweb/quic/la/?p=platform/packages/apps/Browser.git;a=commit;h=9d41dd87506313444124d55f92057ed4831a8dbb

--

Manage Slide Transitions incompatibility in Labs

Slide Tab Transitions is not compatible with Quick Controls because they
both use edge touch detection and solve the same problem. When one
setting is enabled, the other should be disabled. I.e. Slide Tab
Transitions must be disabled in the Labs Settings menu before Quick
Controls can be enabled.

CRs-fixed: 354231

author: Pierre-Antoine LaFayette <plafayet@codeaurora.org>
https://www.codeaurora.org/gitweb/quic/la/?p=platform/packages/apps/Browser.git;a=commit;h=b139ae8afea868118d491dfcb106845f498c2847

--

Null check for onScreenSwitched

RealViewSwitcher allows switching between tabs in the Browser by
swiping. The mapping between tabs and RealViewSwitchers internal
'screens'(i.e. tab position) is done at startup. When this setup
was delayed it caused a NullPointerException.

author:  Bjorn Isaksson <bjorn.isaksson@sonymobile.com>
https://www.codeaurora.org/gitweb/quic/la/?p=platform/packages/apps/Browser.git;a=commit;h=d47e1b5bae5dcbf00a82a793ce5fb581bafc8a77

--

Null check for slide tab transitions in BrowserSettings

BrowserSettings can be used without a browser Controller, e.g.
when accessed from system settings. But some null checks for
this case were missing, these have now been added.

author:  Bjorn Isaksson <bjorn.isaksson@sonymobile.com>
https://www.codeaurora.org/gitweb/quic/la/?p=platform/packages/apps/Browser.git;a=commit;h=58e685424457206dbb728b5d79b1977155705e7d

--

Do not reattach tab on orientation change

When the orientation changes, there is no need to detach and reattach
the current tab. This was causing a crash when Flash forces an
orientation change to play fullscreen video. The previous fix stopped
RealViewSwitcher from calling onScreenSwitched if the screen doesn't
change after a scroll, however the Tab needs to know when the scroll
completes regardless of if the Tab has changed so that it can show the
URL bar.

CRs-fixed: 329432

author: Pierre-Antoine LaFayette <plafayet@codeaurora.org>
https://www.codeaurora.org/gitweb/quic/la/?p=platform/packages/apps/Browser.git;a=commit;h=ed15a9bf26cb49cbb7caa432741d930fc798c4b4

--

Change-Id: Id549fb9639ec83e79fbaea9ccee3e6bf62d5edbc