New Tab Animation
In stock Android, opening a new tab and navigating a link is not
distinguisable. This patch visually indicates that a new tab has been
created using an animation. Also fixes issue with keyboard briefly
showing during slide transitions and an issue with
attachTabToContentView being called twice for the same tab.

author: Pierre-Antoine LaFayette <plafayet@codeaurora.org>
https://www.codeaurora.org/gitweb/quic/la/?p=platform/packages/apps/Browser.git;a=commit;h=ee8d5fb94f6a53f67cb1b580d8bc539404a31aea

--

Prevent unnecessary view removal when using Slide Tab Transitions

When using the feature Slide Tab Transitions and a tab was set as
active, the tab view container was always removed and added again
to a RealViewSwitcher. Also, when a WebView is detached from a
window and it is in full screen mode, it dismisses the full screen
mode. This caused a problem when a plugin (e.g. Flash) wanted to
show something in fullscreen.

Now a tab view container will not be removed add added again to
the same RealViewSwitcher.

author: Bjorn Isaksson <bjorn.isaksson@sonymobile.com>
https://www.codeaurora.org/gitweb/quic/la/?p=platform/packages/apps/Browser.git;a=commit;h=f3dc5430da99146d3430e96a8fc7299ef7cc8f25

--

Fix for orientation change crash

The RealViewSwitcher was getting its new tab view added out of bounds
after a landscape rotation occurred. This was because a tab was being
left behind after the orientation change. The code to stop reattachment
on orientation change was a bug fix for a crash that occurred when Flash
video embedded in the Browser went to fullscreen and changed the
orientation to landscape mode. However, the code changed to prevent
unnecessary reattachments, which should already handle this case, making
the original code for orientation change redundant and caused the loss
of a tab after changing orientation.

CRs-fixed: 386305

author: Pierre-Antoine LaFayette <plafayet@codeaurora.org>
https://www.codeaurora.org/gitweb/quic/la/?p=platform/packages/apps/Browser.git;a=commit;h=1f4964385320235e1750a0331f5e3aa70c33c111

--

Fix incorrect tab removal logic in PhoneUi

If a tab that is earlier in the list gets removed, the current index
needs to be adjusted to point to the correct tab.

CRs-fixed: 407153

--

Open in background tab fix

New Tab Animation and Slide Transitions is not handling the open in
background tab case.

Enable Settings->Advanced->Open in Background Tab and long press a link
to test.

Change-Id: Idade7dd1501c0ae8aaf62c36e3d66819aabbd07f