Remove "fitsSystemWindows=true" property from tab layout
This property definition causes in incorrect layout on the tab view
when SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN or
SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION is set in setSystemUIVisibility.
Either this property needs to be removed or a custom implementation
is required for fitSystemWindows so that this view's size can be
updated upon a change in system UI visibility.

Change-Id: I20f0aa54af6667d2def0122f2494b2f732bfa7dc
CRs-Fixed: 426333