Remove "fitsSystemWindows=true" property from tab layout
This property definition causes in incorrect layout on the tab view
when SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN or
SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION is set in setSystemUIVisibility.
Either this property needs to be removed or a custom implementation
is required for fitSystemWindows so that this view's size can be
updated upon a change in system UI visibility.

CRs-Fixed: 426333
(cherry picked from commit 4666856a993200ab7f83ecddcf9611f0e1569818)

Change-Id: I85313e3dd63e8571a490fe2c1d00f401dafa7c3f