Browser:Progress bar stopped in 5% when failed to load page
When browser opens a page with all data connections disabled,
it fails because of no connection, then it loads a local
error page. In this case the browser actually trys to load two
pages,the original web page and the local error page.
This makes special status transitions.
The common status transitions are
onPageStarted -> onPageFinished -> onProgressChanged(100)
In this case the special status transitions are onPageStarted
->onProgressChanged(100)->onPageStarted->onPageFinished.

Change-Id: Ibdb4605171f18ecfed289a4aa271b21b752d78d1
CRs-Fixed: 387629