Browser: fix cannot search music with yahoo/bing engine
When choose google search engine, this issue does not exist.
But when choose yahoo/bing/baidu search engine and search music
in Music Application, will not show normally but back to Music
App automatically for the same BrowserActivity instance is
destroyed after onNewIntent.

For this situation, will not finish this activity, just keep it.

CRs-Fixed: 494962

Change-Id: Ic53fa5814ab5596bd0bc630d6f3c8771931a8d80