Browser: Fix NPE when save tabs' view state
When browser try to saveViewState,one parameter of the method is
ContentValues.And the contentValues come from createSnapshotValues
method.But the value will be null if the main webview of this tab be
back to background.So NPE happens.

In the monkey test,the main webview of this tab may be back to
background.So we should check whether the values is null in
saveViewState method for safe.

CRs-Fixed: 543716

Change-Id: Ie61708968441c0f5041f45a0f5ada64161fe9c63