Browser: The process bar still display as 5% when try to load a URL
When browser opens a page with all data connections disabled, it fails
because of no connection, then it loads a local error page.
In this case the browser actually trys to load two pages, the original
web page and the local error page. This makes special status transitions.
The common status transitions are onPageStarted -> onPageFinished ->
onProgressChanged(100). In this case the special status transitions are
onPageStarted -> onProgressChanged(100) -> onPageStarted ->
onPageFinished.Through onPageFinished has been done, the mInPageLoad
status is not right.

Change the status of mInPageLoad in onPageFinished function.

CRs-Fixed: 534478

Change-Id: I6a06fa9e15594e5e8ac512f3afde757441a1c8fd