Browser: the bookmarks interface will flicker when the web page is loading
The CursorLoader will help us deal with the cursor observer and
auto-requery, and then through the "onLoadFinished()" returns
the new cursor. Due to the cursor register to watch a content URI
for changes, so any change of URI will cause the cursor observer
action, further cause the UI to redraw.

Replace the URI of cursor observers to monitor , reduce unnecessary
UI updates.

CRs-Fixed: 580369

Change-Id: I9b71b5e44b10a06aa657b5d9436642de1527167f