Browser: fix failed to search song/album/artist
When search engine is default that's google search,
we can succeed to search music by browser. But when
we set engine as no-default, such as baidu, yahoo,
bing, will failed to search and browser will exit
at once. The reason is that we start a new browser
activity instanse if using default engine but still
start previous activity instanse if using no-default,
browser activity will finish itself if can succeed to
handle search intent, so browser activity will be
destroyed when engine is no-default.

We change not to finish browser activity if current
search engine is not default engine.

Change-Id: Iddf46889dad8edda463f6f96ae3142683fa5c493
CRs-Fixed: 619704